rename api class name

+++ uhms-user-service/src/main/java/org/uhms/user/mapper/UserMapper.java
@@ -13,17 +13,17 @@ import org.uhms.user.models.User;
 @Mapper
 public interface UserMapper {
 
-    User getUser(Long id);
+    User getUser(int id);
 
     @Insert("insert into t_user (name,address) values (#{name},#{address})")
     @Options(useGeneratedKeys = true,keyColumn = "id")
-    Integer insertUser(User user);
+    int insertUser(User user);
 
     @Delete("delete from t_user where id = #{id}")
-    Integer deleteUserById(Long id);
+    int deleteUserById(int id);
 
     @Update("update t_user set name=#{name},address=#{address} where id=#{id}")
-    Integer updateUserById(User user);
+    int updateUserById(User user);
 
     @Select("select * from t_user")
     List<User> findAll();
@@ -33,7 +33,7 @@ public interface UserMapper {
     List<User> findUserByName(String name);
 
     @Select("select count(*) from t_user")
-    Long allCount();
+    int allCount();
 
     @Select("select * from t_user limit #{offset},#{limit}")
     List<User> pageUserList(int offset, int limit);
