+++ uhms-user-service/src/main/java/org/uhms/user/services/serviceImpl/UserServiceImpl.java
@@ -38,7 +38,7 @@ public class UserServiceImpl implements IUserService {
     }
 
     @Override
-    public Integer deleteUser(Long id) {
+    public Integer deleteUser(int id) {
         return userMapper.deleteUserById(id);
     }
 }
